requirements.txt
# 网页爬虫：首选 Pyppeteer，备用 Playwright
pyppeteer>=1.0.2
playwright>=1.40.0   # 备用，运行 install-browser 安装 Chromium

# Kimi API（兼容 OpenAI SDK）
openai>=1.0.0

# Word 文档生成
python-docx>=1.1.0

# 环境变量与配置
python-dotenv>=1.0.0

# HTTP 与工具
httpx>=0.27.0
